// src/game/common/interfaces/interval.interface.ts
export interface Iinterval {
    moveLeftAnimation: any;
    moveRightAnimation: any;
    moveRight: any;
    moveLeft: any;
    jump: any;
    fall: any;
}

/*
private moveLeftAnimationInterval: any;
private moveRightAnimationInterval: any;
private moveRightInterval: any;
private moveLeftInterval: any;
private jumpInterval: any;
private fallInterval: any;*/